create css & fixing bug backend

+++ hapus_data.php
@@ -1,8 +1,12 @@
 <?php
     include "koneksi.php";
 
-    $id = $_GET['id'];
-    $delete = mysqli_query($con, "DELETE FROM identitas WHERE id = '$id'");
-
-
-    header('Location: tampilkandata.php?notification=hapus');
+    if (isset($_GET['id'])) {
+        $id = $_GET['id'];
+        $delete = mysqli_query($con, "DELETE FROM identitas WHERE id = '$id'");
+    
+    
+        header('Location: tampilkandata.php?notification=hapus');
+    } else {
+        header('Location: tampilkandata.php');
+    }
